fl-platform/server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
numpy==1.24.3
pyyaml==6.0.1
python-multipart==0.0.6
flwr==1.5.0
torch==2.0.1
pydantic==2.5.0
watchdog==3.0.0